+++ packages/front/svelte/components/Footer.svelte
@@ -27,16 +27,22 @@
                 </p>
                 <ul class="fr-footer__content-list">
                     <li class="fr-footer__content-item">
-                        <a class="fr-footer__content-link" title="Contactez-nous" href="/contact">Contactez-nous</a>
+                        <a class="fr-footer__content-link" target="_blank" href="https://legifrance.gouv.fr">
+                            legifrance.gouv.fr
+                        </a>
                     </li>
                     <li class="fr-footer__content-item">
-                        <a class="fr-footer__content-link" href="https://www.numerique.gouv.fr/">numerique.gouv.fr</a>
+                        <a class="fr-footer__content-link" target="_blank" href="https://www.numerique.gouv.fr/">
+                            numerique.gouv.fr
+                        </a>
                     </li>
                     <li class="fr-footer__content-item">
-                        <a class="fr-footer__content-link" href="https://beta.gouv.fr/">beta.gouv.fr</a>
+                        <a class="fr-footer__content-link" target="_blank" href="https://beta.gouv.fr/">beta.gouv.fr</a>
                     </li>
                     <li class="fr-footer__content-item">
-                        <a class="fr-footer__content-link" href="https://www.gouvernement.fr/">gouvernement.fr</a>
+                        <a class="fr-footer__content-link" target="_blank" href="https://www.gouvernement.fr/">
+                            gouvernement.fr
+                        </a>
                     </li>
                 </ul>
             </div>
@@ -45,6 +51,15 @@
             <ul class="fr-footer__bottom-list">
                 <li class="fr-footer__bottom-item">
                     <a class="fr-footer__bottom-link" href="/accessibilite">Accessibilité : non conforme</a>
+                </li>
+                <li class="fr-footer__bottom-item">
+                    <a
+                        class="fr-footer__bottom-link"
+                        href="https://blog.datasubvention.beta.gouv.fr/statistiques/"
+                        target="_blank"
+                        rel="noreferrer">
+                        Statistiques d'usage
+                    </a>
                 </li>
                 <li class="fr-footer__bottom-item">
                     <a class="fr-footer__bottom-link" href="/mentions-legales">Mentions légales</a>